+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [8845](https://github.com/vegaprotocol/vega/issues/8845) - Add support for network treasury and global insurance accounts.
 - [8846](https://github.com/vegaprotocol/vega/issues/8846) - Add support to transfer recurring transfers to metric based reward
 - [8857](https://github.com/vegaprotocol/vega/issues/8857) - Add a step for getting the balance of the liquidity provider liquidity fee account
+- [8847](https://github.com/vegaprotocol/vega/issues/8847) - Implement internal time trigger data source.
 
 ### 🐛 Fixes